feat(olympus): extract How Olympus works explainer; redirect /architecture -> /system
Move the pipeline explainer out of the /architecture page into a reusable
components/system/how-olympus-works.tsx (title demoted to h2). /architecture now
redirects to /system. Migrate the regression test (Atlas/Hermes phases, real
phase-file tokens, guards against Monthly Synthesis / phase7c_analyst /
phase7cd_debate) onto the component.
